'SameGame' ver 0.02
:sparkles: Title logo
:sparkles: Erasing animation
:sparkles: Game over screen
:+1: Undo
:+1: Sound effects
:bug: Correct judgement of game over
:recycle: Refactoring

+++ samegame/common.cpp
@@ -19,7 +19,7 @@ enum RECORD_STATE_T {
 MyArduboy   arduboy;
 RECORD_T    record;
 int8_t      padX, padY, padRepeatCount;
-uint32_t    lastScore;
+uint16_t    lastScore;
 bool        isInvalid, isRecordDirty;
 
 /*  Local Functions  */
